+++ props-core/src/testFixtures/java/sh/props/textfixtures/TestSource.java
@@ -27,8 +27,8 @@ package sh.props.textfixtures;
 
 import java.util.Map;
 import sh.props.Source;
+import sh.props.SourceFactory;
 import sh.props.annotations.Nullable;
-import sh.props.interfaces.SourceFactory;
 
 /** Creates a source used for tests, which only defines a key=value entry. */
 public class TestSource extends Source {
